requirements.txt
av==12.3.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
coloredlogs==15.0.1
ctranslate2==4.3.1
faster-whisper==1.0.3
filelock==3.15.4
flatbuffers==24.3.25
fsspec==2024.6.1
huggingface-hub==0.24.6
humanfriendly==10.0
idna==3.8
libretranslatepy==2.1.4
mpmath==1.3.0
numpy==2.1.0
onnxruntime==1.19.0
packaging==24.1
piper-phonemize==1.1.0
piper-tts==1.2.0
protobuf==5.27.3
PyAudio==0.2.14
pycparser==2.22
PyYAML==6.0.2
requests==2.32.3
sounddevice==0.5.0
sympy==1.13.2
tokenizers==0.20.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
